Stuff to ensure that we don't re-provision if we don't have to
This commit adds some Puppet code to ensure that we don't reprovision
if we dont' have to.

Puts files in /root/files instead of /tmp so that if /tmp gets cleared
we don't reprovision the file resources.

Package name changed to "jdk" (was "java") since this is what the package
is actually called in dpkg.